ads112c04 on f407 is available

+++ Demo_stm32F407ZGT6/Src/pa_Main/pa_Lvgl/GUIs/MainGUI/MainGUI.h
@@ -5,6 +5,6 @@ namespace GUI
     void initMainGUI();
     void updateRuningTime(int time);
     void updateEncoder(int E1, int E1_d,int E2, int E2_d);
-    void updateAdc(double value);
+    void updateAdc(float value);
 }
 #endif // __MAINGUI_H__
